pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Multi-Agent Deep Research System with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
]

[project.scripts]
research = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
